refactor: move Manhattan/QQ helpers to association module
- move generic plotting helpers out of anoph/phenotypes.py into anoph/association.py\n- add snp_phenotype_association() and association_results() to generate plot-ready inputs\n- wire new mixin into AnophelesDataResource\n- replace tests/anoph/test_phenotypes_plotting.py with tests/anoph/test_association.py
