+++ malariagen_data/anopheles.py
@@ -41,6 +41,7 @@ from .anoph.g123 import AnophelesG123Analysis
 from .anoph.fst import AnophelesFstAnalysis
 from .anoph.h12 import AnophelesH12Analysis
 from .anoph.h1x import AnophelesH1XAnalysis
+from .anoph.association import AnophelesAssociationPlotting
 from .anoph.phenotypes import AnophelesPhenotypeData
 from .mjn import _median_joining_network, _mjn_graph
 from .anoph.hapclust import AnophelesHapClustAnalysis
@@ -97,6 +98,7 @@ class AnophelesDataResource(
     AnophelesGenomeFeaturesData,
     AnophelesGenomeSequenceData,
     AnophelesDescribe,
+    AnophelesAssociationPlotting,
     AnophelesBase,
     AnophelesPhenotypeData,
 ):
